requirements.txt
django
djangorestframework
django-environ
gunicorn
joblib
pandas
numpy
pyarrow
scikit-learn
mlflow
